+++ robot-SHARED/src/sdp/common/Communicator.java
@@ -36,7 +36,7 @@ public interface Communicator {
 	 * @param args the arguments
 	 * @throws IOException in case of error in the connection
 	 */
-	public void sendMessage(opcode op, byte... args) throws IOException;
+	public void sendMessage(opcode op, short... args) throws IOException;
 	
 	/**
 	 * Registers a message listener
